+++ packages/terra-application-header-layout/CHANGELOG.md
@@ -3,6 +3,18 @@ Changelog
 
 Unreleased
 ----------
+### Added
+* Css rule to hide overflow in the case of large content.
+
+2.0.0 - (February 21, 2018)
+------------------
+### Changed
+* Updated to use React 16
+
+1.2.0 - (February 15, 2018)
+------------------
+### Changed
+* Minor dependency updates
 
 1.1.0 - (January 23, 2018)
 ------------------