+++ packages/terra-application-menu-layout/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+2.0.0 - (February 21, 2018)
+------------------
+### Changed
+* Updated to use React 16
+
+1.2.0 - (February 15, 2018)
+------------------
+### Changed
+* Minor dependency updates
+
 1.1.0 - (January 23, 2018)
 ------------------
 ### Changed