+++ packages/terra-application-name/CHANGELOG.md
@@ -3,4 +3,15 @@ ChangeLog
 
 Unreleased
 -----------------
+### Fixed
+* Truncation and padding of the name with new css rules.
+
+2.0.0 - (February 21, 2018)
+------------------
+### Changed
+* Updated to use React 16
+
+1.0.0 - (February 15, 2018)
+------------------
 * Initial stable release
+* Add index file for default main & update default styles.