+++ packages/terra-navigation-layout/CHANGELOG.md
@@ -3,6 +3,14 @@ Changelog
 
 Unreleased
 ----------
+
+2.0.0 - (February 21, 2018)
+------------------
+### Changed
+* Updated to use React 16
+
+1.2.0 - (February 15, 2018)
+------------------
 ###Changed
 * Updating tests to point to new terra-framework-site location.